Fixed raytracing
Changes from another branch

+++ source/resources/Image.cpp
@@ -11,7 +11,7 @@
 std::optional<KS::Image> KS::LoadImageFileFromMemory(const void* filedata, size_t byte_length)
 {
     int height {}, width {}, comp {};
-    auto* stbi_result = stbi_load_from_memory((const stbi_uc*)(filedata), static_cast<int>(byte_length), &width, &height, &comp, 4);
+    auto* stbi_result = stbi_load_from_memory((const stbi_uc*)(filedata), byte_length, &width, &height, &comp, 4);
 
     if (stbi_result != nullptr)
     {
